+++ RELEASE_NOTES.md
@@ -18,18 +18,20 @@ Older versions have their release notes listed in the `Release_Notes` folder.
 
 ## 7.3
 
-This version adds dictation support.
+This version adds a first PREVIEW of dictation support. It's a very early version that is not fully tested, but your feedback is very important so please try it out if you feel like it.
 
 ### ✨ New features
 
 * `Dictation` is a new namespace with types that defines how to perform keyboard dictation.
+* `KeyboardInputViewController` has a `dictationContext` that can be used to manage and observe dictation state.
+* `KeyboardInputViewController` has a `dictationService` that can be used to start dictation from your keyboard.
 * `KeyboardInputViewController` has re-added the old `hostBundleId` property, which can be used to get the ID of the parent app.
 
 ### 👑 Pro changes
 
 * `PreviousAppNavigator` is a new protocol that can be implemented by any type that should be able to navigate back to the previous app.
-* `StandardDictationService` is a new `DictationService` that can be used to perform dictation within an app target.
-* `StandardKeyboardDictationService` is a new `KeyboardDictationService` that can be used to start a dictation operation from a keyboard extension. 
+* `StandardDictationService` is a new service that can be used to perform dictation within an app target.
+* `StandardKeyboardDictationService` is a new service that can be used to start dictation from a keyboard extension. 
 
 
 ## 7.2.1